agent/requirements.txt
google-generativeai==0.7.2
pydantic==2.8.2
PyYAML==6.0.2
python-dotenv==1.0.1
